requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
flake8==3.9.2
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.2.0
pytz==2021.3
sqlparse==0.4.2
